requirements.txt
numpy
opencv-python
scipy
RPi.GPIO
